Demonstrate lifting of 'device specific' info
To enable bootstrapping of Zinc, much information has been put at the
HAL level. An incomplete list of information is:
  - ARM register definition
  - ARM register memory location
  - Vendor register definition
  - Vendor register memory location
  - Register abstraction accessors

The vendor memory locations should be lifted out into a device specific
module, which is ideally generated from a memory map. This simplifies
management of different memory maps and peripheral availability between
devices of the same vendor family.

A selection of these device specific modules should be cultivated in
the contrib directory.