build: allow user to build with dev/debug profile
The user can now invoke builds like the following:

    $ make DEBUG=true EXAMPLE_NAME=blink

This will build artifacts with the dev profile that includes
debug symbols and eliminates optimization that would normally be
included.  This is necessary to get a reasonable debugging experience
in many cases.

A few options were also modified in order to allow for the user
to override them if desired.