Added a repr to generated field enums
Fixes overflowing-literal warnings when using 32-bit constants in 32-bit fields.
Currently using the size of the reg for the size of the enum, but it would also be
reasonable to use the smallest uint that holds all the actual variants, or look at
the format of the literals used to define the variants.