cleanup: remove defunct autotools artifacts
Although it is a little tedious for in-tree builds, we are now able
to reasonably tell users to use cargo with a few extra arguments for
builds.  End user projects are likely to still require a simple
Makefile in order to generate final artifacts.

The configure script and generated Makefile are broken right now.

Signed-off-by: Paul Osborne <osbpau@gmail.com>